Added : Conversion api to get converted value

+++ ModelLayer/Models/Subunit.cs
@@ -15,6 +15,6 @@ namespace ModelLayer
         public String SubunitName { get; set; }
 
         [Required]
-        public float SubUnitsValue { get; set; }
+        public double SubUnitsValue { get; set; }
     }
 }
